+++ src/main/java/de/boondocksulfur/customjukebox/listeners/ParrotDanceListener.java
@@ -25,7 +25,7 @@ public class ParrotDanceListener implements Listener {
         this.plugin = plugin;
     }
 
-    @EventHandler
+    @EventHandler(ignoreCancelled = true)
     public void onDiscInsert(PlayerInteractEvent event) {
         if (!plugin.getConfigManager().isParrotDancingEnabled()) return;